Transferred Swing
It works but the mechanic needs to be adjusted. Should we really be using a joint? Idk.

A few ideas:
Have jumping push you forward a little?
Be able to slide from slightly above the ground (basically generous maneuvers)? Apply the same idea to jumping? Theres a better word for it.
How does slide feel? Jump? Swing? Move?
Use a force based grapple instead of a joint based one? Make it pull you up (maybe by pressing shift)?

Lots of options, easy to implement. Lets get to it!!

A bug:
Sometimes the player will get stuck on corners when walking up stairs or if you approach a platform walking from the right angle.